refactor create hook for real time updates
