+++ components/BlocksTable/index.tsx
@@ -1,11 +1,7 @@
 "use client"
 
-import { useEffect, useMemo, useState } from "react"
-import {
-  keepPreviousData,
-  useQuery,
-  useQueryClient,
-} from "@tanstack/react-query"
+import { useMemo, useState } from "react"
+import { keepPreviousData, useQuery } from "@tanstack/react-query"
 import { PaginationState } from "@tanstack/react-table"
 
 import { Team } from "@/lib/types"
@@ -13,9 +9,9 @@ import { Team } from "@/lib/types"
 import DataTableControlled from "../ui/data-table-controlled"
 
 import { columns } from "./columns"
+import useRealtimeUpdates from "./useRealtimeUpdates"
 
 import { fetchBlocksPaginated, mergeBlocksWithTeams } from "@/lib/blocks"
-import { createClient } from "@/utils/supabase/client"
 
 type Props = {
   className?: string
@@ -40,52 +36,7 @@ const BlocksTable = ({ className, teams }: Props) => {
     placeholderData: keepPreviousData,
   })
 
-  const queryClient = useQueryClient()
-
-  const supabase = createClient()
-
-  useEffect(() => {
-    const blocksChannel = supabase
-      .channel("blocks")
-      .on(
-        "postgres_changes",
-        { event: "INSERT", schema: "public", table: "blocks" },
-        () => {
-          queryClient.invalidateQueries({ queryKey: ["blocks"] })
-        }
-      )
-      .on(
-        "postgres_changes",
-        { event: "UPDATE", schema: "public", table: "blocks" },
-        () => {
-          queryClient.invalidateQueries({ queryKey: ["blocks"] })
-        }
-      )
-      .subscribe()
-
-    const proofsChannel = supabase
-      .channel("proofs")
-      .on(
-        "postgres_changes",
-        { event: "INSERT", schema: "public", table: "proofs" },
-        () => {
-          queryClient.invalidateQueries({ queryKey: ["blocks"] })
-        }
-      )
-      .on(
-        "postgres_changes",
-        { event: "UPDATE", schema: "public", table: "proofs" },
-        () => {
-          queryClient.invalidateQueries({ queryKey: ["blocks"] })
-        }
-      )
-      .subscribe()
-
-    return () => {
-      supabase.removeChannel(blocksChannel)
-      supabase.removeChannel(proofsChannel)
-    }
-  }, [queryClient, supabase])
+  useRealtimeUpdates()
 
   const defaultData = useMemo(() => [], [])
